RB_ Interactive Sky Chart V2
This Version is the update of RB_V1. Like the last version, it is also based on C. Current version of the program can perform all the functions of its last version but with more accuracy, in lessare run time with the help of modified algorithms and with better graphical presentations of charts plus now it can also perform the task to find the direction of any constellation with the help of two functions namely ''where to look' and 'compass'' which are added in it.